chore: convert to use useRoutes
